Tratamento de exceções - exception handling

+++ src/main/java/com/uni/foodstock/services/ProdutoService.java
@@ -1,14 +1,20 @@
 package com.uni.foodstock.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.uni.foodstock.dto.ProdutoDTO;
 import com.uni.foodstock.entidade.Produto;
 import com.uni.foodstock.repositories.ProdutoRepositori;
+import com.uni.foodstock.services.exceptions.DatabaseException;
+import com.uni.foodstock.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProdutoService {
@@ -18,7 +24,8 @@ public class ProdutoService {
 
 	@Transactional(readOnly = true) 						/* Buscar por ID */
 	public ProdutoDTO findById(Long id) {
-		Produto produto = repositori.findById(id).get();
+		Produto produto = repositori.findById(id).orElseThrow(
+				() -> new ResourceNotFoundException("Recurso não encontrado"));
 		return new ProdutoDTO(produto);
 
 	}
@@ -42,18 +49,32 @@ public class ProdutoService {
 
 	@Transactional 											/* Inserir novo produto */
 	public ProdutoDTO update(Long id, ProdutoDTO dto) {
-
-		Produto entidade = repositori.getReferenceById(id);
-		copyDtoToEntity(dto, entidade);
-		entidade = repositori.save(entidade);
-		return new ProdutoDTO(entidade);
+		try {
+			Produto entidade = repositori.getReferenceById(id);
+			copyDtoToEntity(dto, entidade);
+			entidade = repositori.save(entidade);
+			return new ProdutoDTO(entidade);
+		} 
+		catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 
 	}
 	
-	@Transactional 											/* Deletor por ID */
+	@Transactional(propagation = Propagation.SUPPORTS) 		/* Deletor por ID */
 	public void delete(Long id) {
-		repositori.deleteById(id);
+		if (!repositori.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	        	repositori.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DatabaseException("Falha de integridade referencial");
+	   	}
 	}
+
+
 
 	private void copyDtoToEntity(ProdutoDTO dto, Produto entidade) {
 		entidade.setNome(dto.getNome());
